Platforms
Added a static and a movable platform. The movable platform uses the AnimationPlayer2D Node to do this. To mark the position, from the inspectior, it uses the key symbol in the Transform(Node2D) of the scene. The AnimationPlayer2D node is added as a child of the platform scene, not included in the platform scene.