Score
Added the score functionality. Added another label to show how many functions the player has collected. It should be noted that it doesn't uses the CanvasLayer node. So, the score can be seen in a specific place in the world. This step also uses the "Access as Unique Name" function. Also, moved the slime.gd from scenes folder to scripts folder.