Audio
Added audio to the game. The default audio is auto loaded. It should be noted that in the video, the YouTuber uses v4.2.1 and Autoload can be seen in one of the tabs in the Project Settings window. But I'm using v4.5.1 and the Autoload is under the Globals tab. The functionality remains the same. The Coin scene is modified. Audio is added in it. Moreover, an AnimationPlayer is used in this noded. It is used to desappear the coin animation, it collision shape and disappear the scene entirely by changing it's values and after playing the sound, queue_free() is called from the AnimationPlayer to remove this scene. AnimationPlayer > Animation > Animation > New to create a new animaiton.